:recycle: refactor: api prefix 정의와 enum 분리
