+++ src/domainTypes/tobe/chatroom.ts
@@ -1,34 +1,18 @@
-import type { Chat } from "./chat";
-import type { CommonObject } from "./common";
-import type { Court } from "./court";
-import type { User } from "./users";
+import { ChatroomAdminType, ChatroomType } from "@enums/.";
+import type { APIChat } from "./chat";
+import type { APICommon } from "./common";
+import type { APICourt } from "./court";
+import type { APIUser } from "./user";
 
-export interface Chatroom extends CommonObject {
+export interface APIChatroom extends APICommon {
   admins: Admin[];
   type: ChatroomType;
-  court?: Court;
-  participants: User[];
-  lastChat: Chat;
+  court?: APICourt;
+  participants: APIUser[];
+  lastChat: APIChat;
 }
 
-export const chatroomType = {
-  USER: "USER",
-  GROUP: "GROUP",
-  COURT: "COURT",
-} as const;
-
-type ChatroomTypeMap = typeof chatroomType;
-type ChatroomType = ChatroomTypeMap[keyof ChatroomTypeMap];
-
-export const adminType = {
-  OWNER: "OWNER",
-  MAINTAINER: "MAINTAINER",
-} as const;
-
-type AdminTypeMap = typeof adminType;
-type AdminType = AdminTypeMap[keyof AdminTypeMap];
-
 interface Admin {
-  id: Pick<User, "id">;
-  type: AdminType;
+  id: Pick<APIUser, "id">;
+  type: ChatroomAdminType;
 }
