+++ src/domainTypes/tobe/court.ts
@@ -1,6 +1,7 @@
-import type { CommonObject } from "./common";
+import { TextureKey } from "@enums/.";
+import type { APICommon } from "./common";
 
-export interface Court extends CommonObject {
+export interface APICourt extends APICommon {
   name: string;
   latitude: number;
   longitude: number;
@@ -8,17 +9,3 @@ export interface Court extends CommonObject {
   basketCount: number;
   texture: TextureKey;
 }
-
-export const textureType = {
-  RUBBER: "고무",
-  URETHANE: "우레탄",
-  ASPHALT: "아스팔트",
-  SOIL: "흙",
-  CONCRETE: "콘크리트",
-  ETC: "기타",
-} as const;
-
-type TextureTypeMap = typeof textureType;
-
-export type TextureValue = TextureTypeMap[keyof TextureTypeMap];
-export type TextureKey = keyof TextureTypeMap;
