+++ src/domainTypes/tobe/loudspeaker.ts
@@ -1,7 +1,7 @@
-import type { ISOString, OmitAt, SenderObject } from "./common";
-import type { Court } from "./court";
+import type { ISOString, OmitAt, APICommon } from "./common";
+import type { APICourt } from "./court";
 
-export interface Loudspeaker extends SenderObject {
+export interface APILoudspeaker extends APICommon {
   startTime: ISOString;
-  court: OmitAt<Court>;
+  court: OmitAt<APICourt>;
 }
