+++ src/domainTypes/tobe/newCourt.ts
@@ -1,15 +1,6 @@
-import { Court } from "./court";
+import { StatusKey } from "enums/.";
+import { APICourt } from "./court";
 
-export interface NewCourt extends Court {
+export interface APINewCourt extends APICourt {
   status: StatusKey;
 }
-
-export const statusType = {
-  ACCEPT: "승인",
-  DENY: "거절",
-  READY: "대기",
-} as const;
-
-type StatusTypeMap = typeof statusType;
-export type StatusValue = StatusTypeMap[keyof StatusTypeMap];
-export type StatusKey = keyof StatusTypeMap;
