+++ src/domainTypes/tobe/notification.ts
@@ -1,20 +1,12 @@
-import type { CommonObject, OmitAt } from "./common";
-import type { Follow } from "./follow";
-import type { Loudspeaker } from "./loudspeaker";
+import { NotificationType } from "@enums/.";
+import type { APICommon, OmitAt } from "./common";
+import type { APIFollow } from "./follow";
+import type { APILoudspeaker } from "./loudspeaker";
 
-export interface Notification extends CommonObject {
+export interface APINotification extends APICommon {
   type: NotificationType;
-  follow?: OmitAt<Follow>;
-  loudspeaker?: OmitAt<Loudspeaker>;
+  follow?: OmitAt<APIFollow>;
+  loudspeaker?: OmitAt<APILoudspeaker>;
   isRead: boolean;
   isClicked: boolean;
 }
-
-export const notificationType = {
-  FOLLOWING: "FOLLOWING",
-  LOUDSPEAKER: "LOUDSPEAKER",
-} as const;
-
-type NotificationTypeMap = typeof notificationType;
-
-export type NotificationType = NotificationTypeMap[keyof NotificationTypeMap];
