+++ src/domainTypes/tobe/reservation.ts
@@ -1,9 +1,9 @@
-import { CommonObject, OmitAt } from "./common";
-import { Court } from "./court";
+import { APICommon, OmitAt } from "./common";
+import { APICourt } from "./court";
 
-export interface Reservation extends CommonObject {
+export interface APIReservation extends APICommon {
   numberOfReservations: number;
   startTime: string;
   endTime: string;
-  court: OmitAt<Court>;
+  court: OmitAt<APICourt>;
 }
