+++ src/enums/roleType.ts
@@ -1,0 +1,7 @@
+export const roleType = {
+  USER: "USER",
+  ADMIN: "ADMIN",
+} as const;
+
+type RoleMap = typeof roleType;
+export type Role = RoleMap[keyof RoleMap];
